+++ server/incore-common/src/main/java/edu/illinois/ncsa/incore/common/models/Space.java
@@ -16,9 +16,9 @@ import dev.morphia.annotations.Id;
 import dev.morphia.annotations.Property;
 import edu.illinois.ncsa.incore.common.auth.PrivilegeLevel;
 import edu.illinois.ncsa.incore.common.auth.Privileges;
+import jakarta.xml.bind.annotation.XmlRootElement;
 import org.bson.types.ObjectId;
 
-import javax.xml.bind.annotation.XmlRootElement;
 import java.util.ArrayList;
 import java.util.List;